+++ Source/CesiumRuntime/Private/Cesium3DTileset.cpp
@@ -20,6 +20,7 @@
 #include "CesiumRasterOverlay.h"
 #include "CesiumRuntime.h"
 #include "CesiumTransforms.h"
+#include "CreateModelOptions.h"
 #include "Engine/GameViewportClient.h"
 #include "Engine/Texture2D.h"
 #include "Engine/World.h"
@@ -137,6 +138,13 @@ void ACesium3DTileset::SetIonAccessToken(FString InAccessToken) {
     if (this->TilesetSource == ETilesetSource::FromCesiumIon) {
       this->MarkTilesetDirty();
     }
+  }
+}
+
+void ACesium3DTileset::SetAlwaysIncludeTangents(bool bAlwaysIncludeTangents) {
+  if (this->AlwaysIncludeTangents != bAlwaysIncludeTangents) {
+    this->AlwaysIncludeTangents = bAlwaysIncludeTangents;
+    this->MarkTilesetDirty();
   }
 }
 
@@ -428,21 +436,23 @@ public:
 
   virtual void* prepareInLoadThread(
       const CesiumGltf::Model& model,
-      const glm::dmat4& transform) {
-    std::unique_ptr<UCesiumGltfComponent::HalfConstructed> pHalf =
-        UCesiumGltfComponent::CreateOffGameThread(
-            model,
-            transform
+      const glm::dmat4& transform) override {
+
+    CreateModelOptions options;
+    options.alwaysIncludeTangents = this->_pActor->GetAlwaysIncludeTangents();
+
 #if PHYSICS_INTERFACE_PHYSX
-            ,
-            this->_pPhysXCooking
+    options.pPhysXCooking = this->_pPhysXCooking;
 #endif
-        );
+
+    std::unique_ptr<UCesiumGltfComponent::HalfConstructed> pHalf =
+        UCesiumGltfComponent::CreateOffGameThread(model, transform, options);
     return pHalf.release();
   }
 
-  virtual void*
-  prepareInMainThread(Cesium3DTiles::Tile& tile, void* pLoadThreadResult) {
+  virtual void* prepareInMainThread(
+      Cesium3DTiles::Tile& tile,
+      void* pLoadThreadResult) override {
     const Cesium3DTiles::TileContentLoadResult* pContent = tile.getContent();
     if (pContent && pContent->model) {
       std::unique_ptr<UCesiumGltfComponent::HalfConstructed> pHalf(
@@ -463,7 +473,7 @@ public:
   virtual void free(
       Cesium3DTiles::Tile& tile,
       void* pLoadThreadResult,
-      void* pMainThreadResult) noexcept {
+      void* pMainThreadResult) noexcept override {
     if (pLoadThreadResult) {
       UCesiumGltfComponent::HalfConstructed* pHalf =
           reinterpret_cast<UCesiumGltfComponent::HalfConstructed*>(
@@ -483,7 +493,7 @@ public:
 
   virtual void* prepareRasterInMainThread(
       const Cesium3DTiles::RasterOverlayTile& rasterTile,
-      void* pLoadThreadResult) {
+      void* pLoadThreadResult) override {
     const CesiumGltf::ImageCesium& image = rasterTile.getImage();
     if (image.width <= 0 || image.height <= 0) {
       return nullptr;
@@ -527,7 +537,7 @@ public:
       void* pMainThreadRendererResources,
       const CesiumGeometry::Rectangle& textureCoordinateRectangle,
       const glm::dvec2& translation,
-      const glm::dvec2& scale) {
+      const glm::dvec2& scale) override {
     const Cesium3DTiles::TileContentLoadResult* pContent = tile.getContent();
     if (pContent && pContent->model) {
       UCesiumGltfComponent* pGltfContent =
@@ -542,8 +552,6 @@ public:
             scale);
       }
     }
-    // UE_LOG(LogCesium, VeryVerbose, TEXT("No content for attaching raster to
-    // tile"));
   }
 
   virtual void detachRasterInMainThread(
@@ -565,32 +573,9 @@ public:
             textureCoordinateRectangle);
       }
     }
-    // UE_LOG(LogCesium, VeryVerbose, TEXT("No content for detaching raster
-    // from tile"));
   }
 
 private:
-  // static int32 GetObjReferenceCount(UObject* Obj, TArray<UObject*>*
-  // OutReferredToObjects = nullptr)
-  // {
-  // 	if (!Obj ||
-  // 		!Obj->IsValidLowLevelFast())
-  // 	{
-  // 		return -1;
-  // 	}
-
-  // 	TArray<UObject*> ReferredToObjects;             //req outer, ignore
-  // archetype, recursive, ignore transient 	FReferenceFinder
-  // ObjectReferenceCollector(ReferredToObjects, nullptr, false, true, true,
-  // false); 	ObjectReferenceCollector.FindReferences(Obj);
-
-  // 	if (OutReferredToObjects)
-  // 	{
-  // 		OutReferredToObjects->Append(ReferredToObjects);
-  // 	}
-  // 	return ReferredToObjects.Num();
-  // }
-
   void destroyRecursively(USceneComponent* pComponent) {
 
     UE_LOG(
@@ -601,12 +586,6 @@ private:
     if (!pComponent) {
       return;
     }
-
-    // FString newName = TEXT("Destroyed_") + pComponent->GetName();
-    // pComponent->Rename(*newName);
-
-    // TArray<UObject*> before;
-    // GetObjReferenceCount(pComponent, &before);
 
     if (pComponent->IsRegistered()) {
       pComponent->UnregisterComponent();
@@ -619,12 +598,6 @@ private:
 
     pComponent->DestroyPhysicsState();
     pComponent->DestroyComponent();
-
-    // TArray<UObject*> after;
-    // GetObjReferenceCount(pComponent, &after);
-
-    // UE_LOG(LogActor, Warning, TEXT("Before %d, After %d"), before.Num(),
-    // after.Num());
 
     UE_LOG(LogCesium, VeryVerbose, TEXT("Destroying scene component done"));
   }
@@ -665,6 +638,8 @@ void ACesium3DTileset::LoadTileset() {
           std::make_shared<CesiumAsync::SqliteCache>(
               spdlog::default_logger(),
               getCacheDatabaseName()));
+  static CesiumAsync::AsyncSystem asyncSystem(
+      std::make_shared<UnrealTaskProcessor>());
 
   Cesium3DTiles::Tileset* pTileset = this->_pTileset;
 
@@ -693,7 +668,7 @@ void ACesium3DTileset::LoadTileset() {
   Cesium3DTiles::TilesetExternals externals{
       pAssetAccessor,
       std::make_shared<UnrealResourcePreparer>(this),
-      std::make_shared<UnrealTaskProcessor>(),
+      asyncSystem,
       this->CreditSystem ? this->CreditSystem->GetExternalCreditSystem()
                          : nullptr,
       spdlog::default_logger()};
@@ -1249,6 +1224,8 @@ void ACesium3DTileset::PostEditChangeProperty(
       PropName == GET_MEMBER_NAME_CHECKED(ACesium3DTileset, Url) ||
       PropName == GET_MEMBER_NAME_CHECKED(ACesium3DTileset, IonAssetID) ||
       PropName == GET_MEMBER_NAME_CHECKED(ACesium3DTileset, IonAccessToken) ||
+      PropName ==
+          GET_MEMBER_NAME_CHECKED(ACesium3DTileset, AlwaysIncludeTangents) ||
       PropName == GET_MEMBER_NAME_CHECKED(ACesium3DTileset, EnableWaterMask) ||
       PropName == GET_MEMBER_NAME_CHECKED(ACesium3DTileset, Material) ||
       PropName == GET_MEMBER_NAME_CHECKED(ACesium3DTileset, WaterMaterial) ||